qcacmn: Update set pcl handling for channel frequency
Due to channel number ambiguity with introduction of 6Ghz support
policy manager APIs are updated to use frequency values instead
of channel number. Update set pcl WMI handling to adapt for frequency
usage.

CRs-fixed: 2540002

+++ wmi/inc/wmi_unified_sta_param.h
@@ -276,9 +276,9 @@ struct flashing_req_params {
 
 /**
  * struct wmi_pcl_chan_weights - Params to get the valid weighed list
- * @pcl_list: Preferred channel list already sorted in the order of preference
+ * @pcl_list: channel freq list sorted in preferred order
  * @pcl_len: Length of the PCL
- * @saved_chan_list: Valid channel list updated as part of
+ * @saved_chan_list: Valid channel freq list updated as part of
  * WMA_UPDATE_CHAN_LIST_REQ
  * @saved_num_chan: Length of the valid channel list
  * @weighed_valid_list: Weights of the valid channel list. This will have one
@@ -287,9 +287,9 @@ struct flashing_req_params {
  * @weight_list: Weights assigned by policy manager
  */
 struct wmi_pcl_chan_weights {
-	uint8_t pcl_list[MAX_NUM_CHAN];
+	uint32_t pcl_list[MAX_NUM_CHAN];
 	uint32_t pcl_len;
-	uint8_t saved_chan_list[MAX_NUM_CHAN];
+	uint32_t saved_chan_list[MAX_NUM_CHAN];
 	uint32_t saved_num_chan;
 	uint8_t weighed_valid_list[MAX_NUM_CHAN];
 	uint8_t weight_list[MAX_NUM_CHAN];
